Ver 1.4: Complete GPIO Driver and Interrupt Button Implementation Application

+++ STM32F407xx_Drivers/drivers/Inc/gpio_driver.h
@@ -198,7 +198,7 @@ void GPIO_ToggleOutputPin(GPIO_Reg_t* pGPIOx, uint32_t pinNumber);
  * GPIO Interrupt Configuration and Handling
  */
 void GPIO_IRQITConfig(uint8_t IRQNumber, uint8_t EnOrDi);
-void GPIO_IRQPriorityConfig(uint8_t IRQNumber, uint8_t IRQPriorityValue);
+void GPIO_IRQPriorityConfig(uint8_t IRQNumber, uint32_t IRQPriorityValue);
 void GPIO_IRQHandling(uint8_t pinNumber);
 
 #endif /* INC_GPIO_DRIVER_H_ */
